Move post registration actions to RegistrationHandler
